varregion: plug in TerrainData class and modify TerrainModule and LLClientView to use same. This passes a terrain info class around rather than passing a one dimensional array thus allowing variable regions. Update the database storage for variable region sizes. This should be downward compatible (same format for 256x256 regions).

+++ OpenSim/Framework/Constants.cs
@@ -40,7 +40,7 @@ namespace OpenSim.Framework
         public const float TerrainCompression = 100.0f;
         // Since terrain is stored in 16x16 heights, regions must be a multiple of this number and that is the minimum
         public const int MinRegionSize = 16;
-        public const byte TerrainPatchSize = 16;
+        public const int TerrainPatchSize = 16;
 
         public const string DefaultTexture = "89556747-24cb-43ed-920b-47caed15465f";
